Conflicts:
	GamesFleadh/GamesFleadh.sln
	GamesFleadh/Library/AssetServerCacheV3
	GamesFleadh/Library/CurrentLayout.dwlt
	GamesFleadh/Library/CurrentMaximizeLayout.dwlt
	GamesFleadh/Library/InspectorExpandedItems.asset
	GamesFleadh/Library/ProjectSettings.asset
	GamesFleadh/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll
	GamesFleadh/Library/ScriptAssemblies/Assembly-CSharp-Editor.dll.mdb
	GamesFleadh/Library/ScriptAssemblies/Assembly-CSharp.dll
	GamesFleadh/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
	GamesFleadh/Library/assetDatabase3
	GamesFleadh/Library/expandedItems
	GamesFleadh/Library/metadata/00/00000000000000004000000000000000
	GamesFleadh/Library/metadata/3d/3dfeef214c80c084d927ebbaf899f03f
	GamesFleadh/Library/metadata/65/65bb2ec933ef7254c873e46ea6271af9
	GamesFleadh/Library/shadercompiler-32bit-1.log
	GamesFleadh/Library/shadercompiler-64bit-1.log
	GamesFleadh/ProjectSettings/ProjectSettings.asset
	GamesFleadh/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache

+++ GamesFleadh/Assets/Adrians/PlanetStats.cs
@@ -5,7 +5,7 @@ public class PlanetStats : MonoBehaviour {
 
     public string planetName = null;
     public int offence = 0;
-    public int defence = 0;
+    public float defence = 0;
     public int utility = 0;
     public int cost = 20;
     public bool colonised = true;
